+++ src/com/ansorgit/plugins/bash/lang/parser/arithmetic/ArithmeticFactory.java
@@ -29,26 +29,26 @@ public class ArithmeticFactory implements BashTokenTypes, BashElementTypes {
     private ArithmeticFactory() {
     }
 
-    private static ArithmeticParsingFunction simpleExpression = new SimpleArithmeticExpr();
-    private static ArithmeticParsingFunction postIncrement = new PostIncrementExpr(simpleExpression);
-    private static ArithmeticParsingFunction preIncrement = new PreIncrementExpr(postIncrement);
-    private static ArithmeticParsingFunction negation = prefixRepeated(preIncrement, arithmeticNegationOps, ARITH_NEGATION_ELEMENT, "negation");
-    private static ArithmeticParsingFunction exponent = repeated(negation, ARITH_EXPONENT, ARITH_EXPONENT_ELEMENT, "exponent");
-    private static ArithmeticParsingFunction multiplication = repeated(exponent, arithmeticProduct, ARITH_MULTIPLICACTION_ELEMENT, "mulitplication");
-    private static ArithmeticParsingFunction addition = repeated(multiplication, arithmeticAdditionOps, ARITH_SUM_ELEMENT, "addition");
-    private static ArithmeticParsingFunction shift = repeated(addition, arithmeticShiftOps, ARITH_SHIFT_ELEMENT, "shift");
-    private static ArithmeticParsingFunction compoundComparision = repeated(shift, arithmeticCmpOp, ARITH_COMPUND_COMPARISION_ELEMENT, "compoundComparision");
-    private static ArithmeticParsingFunction equality = repeated(compoundComparision, arithmeticEqualityOps, ARITH_EQUALITY_ELEMENT, "equality");
-    private static ArithmeticParsingFunction bitwiseAnd = repeated(equality, ARITH_BITWISE_AND, ARITH_BIT_AND_ELEMENT, "bitwiseAnd");
-    private static ArithmeticParsingFunction bitwiseXor = repeated(bitwiseAnd, ARITH_BITWISE_XOR, ARITH_BIT_XOR_ELEMENT, "bitwiseXor");
-    private static ArithmeticParsingFunction bitwiseOr = repeated(bitwiseXor, PIPE, ARITH_BIT_OR_ELEMENT, "bitwiseOr");
-    private static ArithmeticParsingFunction logicalAnd = repeated(bitwiseOr, AND_AND, ARITH_LOGIC_AND_ELEMENT, "bitwiseAnd");
-    private static ArithmeticParsingFunction logicalOr = repeated(logicalAnd, OR_OR, ARITH_LOGIC_OR_ELEMENT, "logicalOr");
-    private static ArithmeticParsingFunction variableOperator = repeated(logicalOr, VARIABLE, ARITH_VARIABLE_OPERATOR_ELEMENT, "varOperator"); //math with a variable as operator
-    private static TernaryExpression ternary = new TernaryExpression(variableOperator);
-    private static ArithmeticParsingFunction simpleAssignment = new AbstractAssignment(ternary, TokenSet.create(EQ));
-    private static ArithmeticParsingFunction assignmentCombination = new AbstractAssignment(simpleAssignment, arithmeticAssign);
-    private static ArithmeticParsingFunction assignmentChain = repeated(assignmentCombination, COMMA, ARITH_ASSIGNMENT_CHAIN_ELEMENT, "assignmentChain");
+    private static final ArithmeticParsingFunction simpleExpression = new SimpleArithmeticExpr();
+    private static final ArithmeticParsingFunction postIncrement = new PostIncrementExpr(simpleExpression);
+    private static final ArithmeticParsingFunction preIncrement = new PreIncrementExpr(postIncrement);
+    private static final ArithmeticParsingFunction negation = prefixRepeated(preIncrement, arithmeticNegationOps, ARITH_NEGATION_ELEMENT, "negation");
+    private static final ArithmeticParsingFunction exponent = repeated(negation, ARITH_EXPONENT, ARITH_EXPONENT_ELEMENT, "exponent");
+    private static final ArithmeticParsingFunction multiplication = repeated(exponent, arithmeticProduct, ARITH_MULTIPLICACTION_ELEMENT, "mulitplication");
+    private static final ArithmeticParsingFunction addition = repeated(multiplication, arithmeticAdditionOps, ARITH_SUM_ELEMENT, "addition");
+    private static final ArithmeticParsingFunction shift = repeated(addition, arithmeticShiftOps, ARITH_SHIFT_ELEMENT, "shift");
+    private static final ArithmeticParsingFunction compoundComparision = repeated(shift, arithmeticCmpOp, ARITH_COMPUND_COMPARISION_ELEMENT, "compoundComparision");
+    private static final ArithmeticParsingFunction equality = repeated(compoundComparision, arithmeticEqualityOps, ARITH_EQUALITY_ELEMENT, "equality");
+    private static final ArithmeticParsingFunction bitwiseAnd = repeated(equality, ARITH_BITWISE_AND, ARITH_BIT_AND_ELEMENT, "bitwiseAnd");
+    private static final ArithmeticParsingFunction bitwiseXor = repeated(bitwiseAnd, ARITH_BITWISE_XOR, ARITH_BIT_XOR_ELEMENT, "bitwiseXor");
+    private static final ArithmeticParsingFunction bitwiseOr = repeated(bitwiseXor, PIPE, ARITH_BIT_OR_ELEMENT, "bitwiseOr");
+    private static final ArithmeticParsingFunction logicalAnd = repeated(bitwiseOr, AND_AND, ARITH_LOGIC_AND_ELEMENT, "bitwiseAnd");
+    private static final ArithmeticParsingFunction logicalOr = repeated(logicalAnd, OR_OR, ARITH_LOGIC_OR_ELEMENT, "logicalOr");
+    private static final ArithmeticParsingFunction variableOperator = repeated(logicalOr, VARIABLE, ARITH_VARIABLE_OPERATOR_ELEMENT, "varOperator"); //math with a variable as operator
+    private static final TernaryExpression ternary = new TernaryExpression(variableOperator);
+    private static final ArithmeticParsingFunction simpleAssignment = new AbstractAssignment(ternary, TokenSet.create(EQ));
+    private static final ArithmeticParsingFunction assignmentCombination = new AbstractAssignment(simpleAssignment, arithmeticAssign);
+    private static final ArithmeticParsingFunction assignmentChain = repeated(assignmentCombination, COMMA, ARITH_ASSIGNMENT_CHAIN_ELEMENT, "assignmentChain");
 
     public static ArithmeticParsingFunction entryPoint() {
         return assignmentChain;