+++ Changelog.md
@@ -1,3 +1,6 @@
+#### 2019-02-08:
+ - [#657](https://github.com/BashSupport/BashSupport/issues/657): Exception when the arithmetic operator >> was used
+
 #### 2018-12-15:
  - [#645](https://github.com/BashSupport/BashSupport/issues/645): fix NPE in keyword handling
  - [#634](https://github.com/BashSupport/BashSupport/issues/634): Handle invalid paths in include file inspection